requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
databases==0.9.0
docker==7.1.0
et_xmlfile==2.0.0
fastapi==0.116.1
fonttools==4.59.0
fpdf==1.7.2
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==10.4.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
reportlab==4.4.3
requests==2.32.5
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
tkcalendar==1.6.1
ttkbootstrap==1.14.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
